reorganization of code; fix vscode code problems; make npm run build work;
